requirements.txt
streamlit
yt-dlp
openai-whisper
python-dotenv
torch
faster-whisper
faiss-cpu
groq
sentence-transformers
reportlab
numpy